fix(aiff,wav): compare encoded size as uint64 for 32-bit builds
int(^uint32(0)) overflowed at compile time on 32-bit targets such as ARMv6. Comparing both lengths as uint64 keeps the 4 GiB limit on 64-bit and lets 32-bit builds compile.

+++ aiff/aiff.go
@@ -302,7 +302,7 @@ func (f *File) encode() ([]byte, error) {
 			inner.WriteByte(0)
 		}
 	}
-	if inner.Len() > int(^uint32(0)) {
+	if uint64(inner.Len()) > uint64(^uint32(0)) {
 		return nil, errors.New("aiff: encoded body exceeds 4 GiB")
 	}
 	var out bytes.Buffer
